+++ CHANGELOG.md
@@ -3443,3 +3443,4 @@
 - [2026-08-07 10:37] perf(analytics): optimize memory consumption in analytics list viewport
 - [2026-08-07 11:38] perf(storage): apply RepaintBoundary on animated storage widgets
 - [2026-08-07 11:47] test(auth): add unit tests for auth business logic
+- [2026-08-07 16:25] feat(auth): implement shimmer loading skeleton in auth